fix: match `color.RGBA()` behavior from the standard library (#30)
* Update whisker in flake

The template requires whisker 2.5.1 but in nix we were stuck with 2.3.0

* Make alpha 255 by default

I think this was the intended behavior, we want the colors to be fully
opaque.

* Match the expected behavior of color.RGBA() and update tests

In the standard library color.RGBA() expects returned values to be in
the interval [0, 0xffff]. With this patch we match the behavior of
image.color.RGBA.RGBA().

Tests were updated accordingly and a test was added to make sure we
match the implementation of the standard library.

+++ mocha.go
@@ -13,7 +13,7 @@ func (mocha) Name() string { return "mocha" }
 func (mocha) Rosewater() Color {
 	return Color{
 		Hex: "#f5e0dc",
-		RGB: [3]uint32{245, 224, 220},
+		RGB: [3]uint8{245, 224, 220},
 		HSL: [3]float32{10, 0.56, 0.91},
 	}
 }
@@ -22,7 +22,7 @@ func (mocha) Rosewater() Color {
 func (mocha) Flamingo() Color {
 	return Color{
 		Hex: "#f2cdcd",
-		RGB: [3]uint32{242, 205, 205},
+		RGB: [3]uint8{242, 205, 205},
 		HSL: [3]float32{0, 0.59, 0.88},
 	}
 }
@@ -31,7 +31,7 @@ func (mocha) Flamingo() Color {
 func (mocha) Pink() Color {
 	return Color{
 		Hex: "#f5c2e7",
-		RGB: [3]uint32{245, 194, 231},
+		RGB: [3]uint8{245, 194, 231},
 		HSL: [3]float32{316, 0.72, 0.86},
 	}
 }
@@ -40,7 +40,7 @@ func (mocha) Pink() Color {
 func (mocha) Mauve() Color {
 	return Color{
 		Hex: "#cba6f7",
-		RGB: [3]uint32{203, 166, 247},
+		RGB: [3]uint8{203, 166, 247},
 		HSL: [3]float32{267, 0.84, 0.81},
 	}
 }
@@ -49,7 +49,7 @@ func (mocha) Mauve() Color {
 func (mocha) Red() Color {
 	return Color{
 		Hex: "#f38ba8",
-		RGB: [3]uint32{243, 139, 168},
+		RGB: [3]uint8{243, 139, 168},
 		HSL: [3]float32{343, 0.81, 0.75},
 	}
 }
@@ -58,7 +58,7 @@ func (mocha) Red() Color {
 func (mocha) Maroon() Color {
 	return Color{
 		Hex: "#eba0ac",
-		RGB: [3]uint32{235, 160, 172},
+		RGB: [3]uint8{235, 160, 172},
 		HSL: [3]float32{350, 0.65, 0.77},
 	}
 }
@@ -67,7 +67,7 @@ func (mocha) Maroon() Color {
 func (mocha) Peach() Color {
 	return Color{
 		Hex: "#fab387",
-		RGB: [3]uint32{250, 179, 135},
+		RGB: [3]uint8{250, 179, 135},
 		HSL: [3]float32{23, 0.92, 0.75},
 	}
 }
@@ -76,7 +76,7 @@ func (mocha) Peach() Color {
 func (mocha) Yellow() Color {
 	return Color{
 		Hex: "#f9e2af",
-		RGB: [3]uint32{249, 226, 175},
+		RGB: [3]uint8{249, 226, 175},
 		HSL: [3]float32{41, 0.86, 0.83},
 	}
 }
@@ -85,7 +85,7 @@ func (mocha) Yellow() Color {
 func (mocha) Green() Color {
 	return Color{
 		Hex: "#a6e3a1",
-		RGB: [3]uint32{166, 227, 161},
+		RGB: [3]uint8{166, 227, 161},
 		HSL: [3]float32{115, 0.54, 0.76},
 	}
 }
@@ -94,7 +94,7 @@ func (mocha) Green() Color {
 func (mocha) Teal() Color {
 	return Color{
 		Hex: "#94e2d5",
-		RGB: [3]uint32{148, 226, 213},
+		RGB: [3]uint8{148, 226, 213},
 		HSL: [3]float32{170, 0.57, 0.73},
 	}
 }
@@ -103,7 +103,7 @@ func (mocha) Teal() Color {
 func (mocha) Sky() Color {
 	return Color{
 		Hex: "#89dceb",
-		RGB: [3]uint32{137, 220, 235},
+		RGB: [3]uint8{137, 220, 235},
 		HSL: [3]float32{189, 0.71, 0.73},
 	}
 }
@@ -112,7 +112,7 @@ func (mocha) Sky() Color {
 func (mocha) Sapphire() Color {
 	return Color{
 		Hex: "#74c7ec",
-		RGB: [3]uint32{116, 199, 236},
+		RGB: [3]uint8{116, 199, 236},
 		HSL: [3]float32{199, 0.76, 0.69},
 	}
 }
@@ -121,7 +121,7 @@ func (mocha) Sapphire() Color {
 func (mocha) Blue() Color {
 	return Color{
 		Hex: "#89b4fa",
-		RGB: [3]uint32{137, 180, 250},
+		RGB: [3]uint8{137, 180, 250},
 		HSL: [3]float32{217, 0.92, 0.76},
 	}
 }
@@ -130,7 +130,7 @@ func (mocha) Blue() Color {
 func (mocha) Lavender() Color {
 	return Color{
 		Hex: "#b4befe",
-		RGB: [3]uint32{180, 190, 254},
+		RGB: [3]uint8{180, 190, 254},
 		HSL: [3]float32{232, 0.97, 0.85},
 	}
 }
@@ -139,7 +139,7 @@ func (mocha) Lavender() Color {
 func (mocha) Text() Color {
 	return Color{
 		Hex: "#cdd6f4",
-		RGB: [3]uint32{205, 214, 244},
+		RGB: [3]uint8{205, 214, 244},
 		HSL: [3]float32{226, 0.64, 0.88},
 	}
 }
@@ -148,7 +148,7 @@ func (mocha) Text() Color {
 func (mocha) Subtext1() Color {
 	return Color{
 		Hex: "#bac2de",
-		RGB: [3]uint32{186, 194, 222},
+		RGB: [3]uint8{186, 194, 222},
 		HSL: [3]float32{227, 0.35, 0.8},
 	}
 }
@@ -157,7 +157,7 @@ func (mocha) Subtext1() Color {
 func (mocha) Subtext0() Color {
 	return Color{
 		Hex: "#a6adc8",
-		RGB: [3]uint32{166, 173, 200},
+		RGB: [3]uint8{166, 173, 200},
 		HSL: [3]float32{228, 0.24, 0.72},
 	}
 }
@@ -166,7 +166,7 @@ func (mocha) Subtext0() Color {
 func (mocha) Overlay2() Color {
 	return Color{
 		Hex: "#9399b2",
-		RGB: [3]uint32{147, 153, 178},
+		RGB: [3]uint8{147, 153, 178},
 		HSL: [3]float32{228, 0.17, 0.64},
 	}
 }
@@ -175,7 +175,7 @@ func (mocha) Overlay2() Color {
 func (mocha) Overlay1() Color {
 	return Color{
 		Hex: "#7f849c",
-		RGB: [3]uint32{127, 132, 156},
+		RGB: [3]uint8{127, 132, 156},
 		HSL: [3]float32{230, 0.13, 0.55},
 	}
 }
@@ -184,7 +184,7 @@ func (mocha) Overlay1() Color {
 func (mocha) Overlay0() Color {
 	return Color{
 		Hex: "#6c7086",
-		RGB: [3]uint32{108, 112, 134},
+		RGB: [3]uint8{108, 112, 134},
 		HSL: [3]float32{231, 0.11, 0.47},
 	}
 }
@@ -193,7 +193,7 @@ func (mocha) Overlay0() Color {
 func (mocha) Surface2() Color {
 	return Color{
 		Hex: "#585b70",
-		RGB: [3]uint32{88, 91, 112},
+		RGB: [3]uint8{88, 91, 112},
 		HSL: [3]float32{233, 0.12, 0.39},
 	}
 }
@@ -202,7 +202,7 @@ func (mocha) Surface2() Color {
 func (mocha) Surface1() Color {
 	return Color{
 		Hex: "#45475a",
-		RGB: [3]uint32{69, 71, 90},
+		RGB: [3]uint8{69, 71, 90},
 		HSL: [3]float32{234, 0.13, 0.31},
 	}
 }
@@ -211,7 +211,7 @@ func (mocha) Surface1() Color {
 func (mocha) Surface0() Color {
 	return Color{
 		Hex: "#313244",
-		RGB: [3]uint32{49, 50, 68},
+		RGB: [3]uint8{49, 50, 68},
 		HSL: [3]float32{237, 0.16, 0.23},
 	}
 }
@@ -220,7 +220,7 @@ func (mocha) Surface0() Color {
 func (mocha) Base() Color {
 	return Color{
 		Hex: "#1e1e2e",
-		RGB: [3]uint32{30, 30, 46},
+		RGB: [3]uint8{30, 30, 46},
 		HSL: [3]float32{240, 0.21, 0.15},
 	}
 }
@@ -229,7 +229,7 @@ func (mocha) Base() Color {
 func (mocha) Mantle() Color {
 	return Color{
 		Hex: "#181825",
-		RGB: [3]uint32{24, 24, 37},
+		RGB: [3]uint8{24, 24, 37},
 		HSL: [3]float32{240, 0.21, 0.12},
 	}
 }
@@ -238,7 +238,7 @@ func (mocha) Mantle() Color {
 func (mocha) Crust() Color {
 	return Color{
 		Hex: "#11111b",
-		RGB: [3]uint32{17, 17, 27},
+		RGB: [3]uint8{17, 17, 27},
 		HSL: [3]float32{240, 0.23, 0.09},
 	}
 }
